requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.16
    # via python-binance
aiosignal==1.3.2
    # via aiohttp
attrs==25.3.0
    # via aiohttp
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
dateparser==1.2.1
    # via python-binance
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via
    #   requests
    #   yarl
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
pycryptodome==3.22.0
    # via python-binance
python-binance==1.0.28
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via dateparser
pytz==2025.2
    # via dateparser
regex==2024.11.6
    # via dateparser
requests==2.32.3
    # via
    #   -r requirements.in
    #   python-binance
six==1.17.0
    # via
    #   python-binance
    #   python-dateutil
tzlocal==5.3.1
    # via dateparser
urllib3==2.4.0
    # via requests
websockets==15.0.1
    # via python-binance
yarl==1.19.0
    # via aiohttp
